Improved python interface and fixes
macro_times, micro_times, and routing_channels
can now be accessed as properties of a TTTR object.

The method get_filename is fixed. Before it was not
returning valid python strings. Now the filename is
set upon initialization to the absolute path of the
file and a std::string is returned.

added method get_tttr_container_type that returns
the container reading method that was used when the TTTR
file was opened.